Issue #479
Milestone B, Confirm box now appears for delete buttons while viewing, while activating an entity like Project/Technician. Currently available only for Projects and Technician sections

+++ Web/js/app/controllers/project_manager.js
@@ -14,30 +14,19 @@ $(document).ready(function() {
       if (key === "edit") {
         project_manager.retrieveProject(options.$trigger);
       } else if (key === "delete") { 
-		var attr = $(this).attr('confirmdelete');
-		if (typeof attr !== typeof undefined && attr !== false) {
-		  bootbox.confirm({
-		    buttons: {
-			  confirm: {
-			    label: 'Ok',
-			    className: 'btn btn-default confirmbuttons'
-			  },
-			  cancel: {
-			    label: 'Cancel',
-			    className: 'btn btn-primary confirmbuttons'
-			  }
-		    },
-		    message: attr,
-		    callback: function(result) {
-			  if(result)
-			  	project_manager.delete(parseInt(options.$trigger.attr("data-project-id")));
-		    }
-		  });
-		}
-		else
-		{
-		  project_manager.delete(parseInt(options.$trigger.attr("data-project-id")));
-		}
+				var msg = $('#confirmmsg-delete').val();
+				if (typeof msg !== typeof undefined && msg !== false) {
+		  		utils.showConfirmBox(msg, function(result){
+						if(result)
+						{
+							project_manager.delete(parseInt(options.$trigger.attr("data-project-id")));
+						}
+				  });
+				}
+				else
+				{
+					project_manager.delete(parseInt(options.$trigger.attr("data-project-id")));
+				}
       }
     },
     items: {
@@ -69,7 +58,19 @@ $(document).ready(function() {
     }
   });
   $(".from-inactive-list").click(function () {
-    project_manager.updateProjects("active", project_ids);
+		var msg = $('#confirmmsg-activate').val();
+		if (typeof msg !== typeof undefined && msg !== false) {
+			utils.showConfirmBox(msg, function(result){
+				if(result)
+				{
+					project_manager.updateProjects("active", project_ids);
+				}
+			});
+		}
+		else
+		{
+			project_manager.updateProjects("active", project_ids);
+		}
   });
   $(".from-active-list").click(function () {
     project_manager.updateProjects("inactive", project_ids);
@@ -100,7 +101,20 @@ $(document).ready(function() {
   });//Edit a project
 
   $("#btn_delete_project").click(function () {
-    project_manager.delete(parseInt(utils.getQueryVariable("project_id")));
+		var msg = $('#confirmmsg-delete').val();
+		if (typeof msg !== typeof undefined && msg !== false) {
+			utils.showConfirmBox(msg, function(result){
+				if(result)
+				{
+					project_manager.delete(parseInt(utils.getQueryVariable("project_id")));
+				}
+			});
+		}
+		else
+		{
+			project_manager.delete(parseInt(utils.getQueryVariable("project_id")));
+		}
+    
   });//Delete a project
 
   if (utils.getQueryVariable("mode") === "edit") {
